fix-agentic/home-fix-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "home-fix-agent"
version = "0.1.0"
description = "Agentic home project assistant — photo to product order"
requires-python = ">=3.11"
dependencies = [
    "pydantic==2.11.1",
    "pyyaml==6.0.2",
    "openai==1.82.0",
    "httpx==0.28.1",
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "python-multipart==0.0.20",
    "jinja2==3.1.6",
    "python-dateutil==2.9.0.post0",
]

[project.optional-dependencies]
dev = ["pytest==8.3.5"]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
